Pin kernal to 6.15.2.
Updating to 6.15.4 results in a system that is so slow as to be
unusable.

So, here, I've pinned the kernel and added a makefile target that let's
me quickly check what version of the kernel is on latest. When latest
contains 6.15.5, I can try it to see if the issue is fixed.